requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
deprecation==2.1.0
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
joblib==1.5.2
numpy==2.3.2
packaging==25.0

# Reddit API
praw==7.8.1
prawcore==2.4.0

# Database
postgrest==1.1.1
psycopg2-binary==2.9.10
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
storage3==0.12.1

# Environment & Config
python-dotenv==1.1.1
PyYAML==6.0.2

# Authentication
PyJWT==2.10.1
pydantic==2.11.7
pydantic_core==2.33.2

# ML & Data Science
scikit-learn==1.7.1
scipy==1.16.1
xgboost==3.0.4
pandas>=2.0.0
torch>=2.0.0
transformers==4.56.0
tokenizers==0.22.0
safetensors==0.6.2

# Data Processing
python-dateutil==2.9.0.post0
regex==2025.8.29
tqdm==4.67.1

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
six==1.17.0
sniffio==1.3.1
StrEnum==0.4.15
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.5.0

# WebSockets
websocket-client==1.8.0
websockets==15.0.1
realtime==2.7.0
requests==2.32.5
